[lldb-dap] Adding type serialization to protocol messages.
This is a work in progress refactor to add explicit types instead of generic 'llvm::json::Value' types to the DAP protocol.

This updates RequestHandler to have take the type of the arguments and response body for serialization for requests and a typed event helper `RegisterEvent` for sending an event.

The updated request is the 'source' request and the updated event is the 'exited' event to make sure the basic structures are working as expected.