[lldb] Prototyping type checking python files when running tests.
This is a prototype of integrating mypy into lldb's test suite to validated the files as part of running tests.

To enable this, add `-DLLDB_PYTHON_TESTS_CHECK_TYPES:STRING=warn|strict` to your cmake configuration.

When this flag is set, prior to executing a test the dotest.py helper will run mypy, which will report any type checking issues to stdout.