[lldb-mcp] Work in progress.
This is a prototype of a coordinator between lldb instances and LLM tools using a new lldb-mcp helper binary.

This will launch the lldb-mcp binary on-demand and the lldb-mcp binary can act as a server for coordinating with an LLM.

This is still a WIP.

I still need to re-orgnaize the MCP protocol into Host and simplify the MCP lldb plugin.

With the new approach I think we can simplify the plugin and let the lldb-mcp server handle more of the other operations.