[lldb] MCP Changes.
WIP, not ready for review yet.

This includes a new abstraction in JSONTransport for creating bindings to methods / events that can be used in lldb-mcp or lldb-dap.

This also includes some adjustements to the lldb-mcp binary. It now auto-connects to the existing lldb instance, if one exists. There is future work to launch an lldb instance or relay to a specific lldb instance.